renamed class to function as an entity object
